Incorporated review comments changes for #1985, #2264
