+++ data-prepper-plugins/opensearch-source/build.gradle
@@ -8,6 +8,7 @@ dependencies {
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.14.2'
     implementation 'software.amazon.awssdk:s3'
     implementation 'software.amazon.awssdk:sts'
+    implementation project(path: ':data-prepper-core')
     testImplementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
 }
