+++ data-prepper-plugins/opensearch-source/src/main/java/org/opensearch/dataprepper/plugins/source/opensearch/configuration/ConnectionConfiguration.java
@@ -5,21 +5,9 @@
 package org.opensearch.dataprepper.plugins.source.opensearch.configuration;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotNull;
 import java.nio.file.Path;
-import java.util.List;
 
 public class ConnectionConfiguration {
-
-  @NotNull
-  @JsonProperty("hosts")
-  private List<String> hosts;
-
-  @JsonProperty("username")
-  private String username;
-
-  @JsonProperty("password")
-  private String password;
 
   @JsonProperty("cert")
   private Path certPath;
@@ -33,18 +21,6 @@ public class ConnectionConfiguration {
   @JsonProperty("insecure")
   private boolean insecure;
 
-  public List<String> getHosts() {
-    return hosts;
-  }
-
-  public String getUsername() {
-    return username;
-  }
-
-  public String getPassword() {
-    return password;
-  }
-
   public Path getCertPath() {
     return certPath;
   }
@@ -57,4 +33,7 @@ public class ConnectionConfiguration {
     return connectTimeout;
   }
 
+  public boolean isInsecure() {
+    return insecure;
+  }
 }
