--- data-prepper-plugins/opensearch-source/src/main/java/org/opensearch/dataprepper/plugins/source/opensearch/configuration/RetryConfiguration.java
@@ -1,18 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.opensearch.dataprepper.plugins.source.opensearch.configuration;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class RetryConfiguration {
-
-  @JsonProperty("max_retries")
-  private Integer maxRetries;
-
-  public Integer getMaxRetries() {
-    return maxRetries;
-  }
-}
