+++ data-prepper-plugins/opensearch-source/src/main/java/org/opensearch/dataprepper/plugins/source/opensearch/configuration/SchedulingParameterConfiguration.java
@@ -8,25 +8,23 @@ package org.opensearch.dataprepper.plugins.source.opensearch.configuration;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.DurationDeserializer;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.DurationSerializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
 import jakarta.validation.constraints.Min;
+import org.opensearch.dataprepper.parser.DataPrepperDurationDeserializer;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
 
 public class SchedulingParameterConfiguration {
 
-    @JsonSerialize(using = DurationSerializer.class)
-    @JsonDeserialize(using = DurationDeserializer.class)
     @JsonProperty("rate")
+    @JsonDeserialize(using = DataPrepperDurationDeserializer.class)
     private Duration rate;
 
     @Min(1)
     @JsonProperty("job_count")
-    private int jobCount;
+    private int jobCount = 1;
 
     @JsonSerialize(using = LocalDateTimeSerializer.class)
     @JsonDeserialize(using = LocalDateTimeDeserializer.class)
